Adjusting to_c impls with the proper type casts

+++ libnss/src/shadow.rs
@@ -16,13 +16,13 @@ impl ToC<CShadow> for Shadow {
     unsafe fn to_c(&self, result: *mut CShadow, buffer: &mut CBuffer) -> std::io::Result<()> {
         (*result).name = buffer.write_str(&self.name)?;
         (*result).passwd = buffer.write_str(&self.passwd)?;
-        (*result).last_change = self.last_change;
-        (*result).change_min_days = self.change_min_days;
-        (*result).change_max_days = self.change_max_days;
-        (*result).change_warn_days = self.change_warn_days;
-        (*result).change_inactive_days = self.change_inactive_days;
-        (*result).expire_date = self.expire_date;
-        (*result).reserved = self.reserved;
+        (*result).last_change = self.last_change as libc::c_long;
+        (*result).change_min_days = self.change_min_days as libc::c_long;
+        (*result).change_max_days = self.change_max_days as libc::c_long;
+        (*result).change_warn_days = self.change_warn_days as libc::c_long;
+        (*result).change_inactive_days = self.change_inactive_days as libc::c_long;
+        (*result).expire_date = self.expire_date as libc::c_long;
+        (*result).reserved = self.reserved as libc::c_ulong;
         Ok(())
     }
 }
